extract pattern matching codes to _common for easier re-use
